pyproject.toml
[project]
name = "saas-blog-api"
version = "0.1.0"
description = "fastapi api for saas-blog site"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=24.1.0",
    "aiosqlite>=0.21.0",
    "alembic>=1.16.4",
    "asgi-lifespan>=2.1.0",
    "asyncpg>=0.30.0",
    "bcrypt==4.0.1",
    "black>=25.1.0",
    "fakeredis[lua]>=2.31.0",
    "fastapi-limiter>=0.1.6",
    "fastapi[standard]>=0.116.1",
    "gunicorn>=23.0.0",
    "httpx>=0.28.1",
    "isort>=6.0.1",
    "nanoid>=2.0.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.10.1",
    "pyjwt>=2.10.1",
    "pyproject-toml>=0.1.0",
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "python-dotenv>=1.1.1",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.20",
    "slugify>=0.0.1",
    "sqlmodel>=0.0.24",
    "sse-starlette>=3.0.2",
    "trio>=0.30.0",
]
